Check the role name on frontend

+++ frontend/src/app/modules/members/members-page/members-page.component.ts
@@ -145,7 +145,11 @@ export class MembersPageComponent extends BaseComponent implements OnInit {
             acceptButton: { class: 'p-button-primary p-button-outlined' },
             cancelButton: { class: 'p-button-secondary p-button-outlined' },
             accept: () => {
-                this.deleteMember(memberItem);
+                if (memberItem.member.role.name != 'Owner') {
+                    this.deleteMember(memberItem);
+                } else {
+                    this.toastNotifications.error('You cannot delete this member');
+                }
             },
         });
     }
